[CELEBORN-1457] Avoid NPE during shuffle data cleanup

+++ worker/src/main/scala/org/apache/celeborn/service/deploy/worker/storage/StorageManager.scala
@@ -574,7 +574,7 @@ final private[worker] class StorageManager(conf: CelebornConf, workerSource: Abs
         }
         if (workerGracefulShutdown) {
           committedFileInfos.remove(shuffleKey)
-          if (cleanDB) {
+          if (cleanDB && null != db) {
             db.delete(dbShuffleKey(shuffleKey))
           }
         }
